test: use pair-wise short-circuit for shared borders
It works for C-shaped + triangle, however the composition with the
pentagon is strange.

Future ideas: not only shared borders, but remove any constraint that
intersects another polygon (interior or boundary) -- this would also
(ideally) reshape the triangle portion near the corners to more
triangular level sets. Implies changing the logic of the shared_borders
array (2d->3d) + inner for loop.